extract note_tree_view.dart
