+++ notes/flutter_web/lib/routes/layout.dart
@@ -65,124 +65,6 @@ class RootLayoutState extends State<RootLayout> {
   }
 }
 
-class _NoteTreeView extends StatefulWidget {
-  const _NoteTreeView({required this.view});
-
-  final Value<String> view;
-
-  @override
-  State<StatefulWidget> createState() {
-    return _NoteTreeViewState();
-  }
-}
-
-class _NoteTreeViewState extends State<_NoteTreeView> {
-  final Value<bool> includeDraft = false.signal();
-  final Value<To?> selected = (null as To?).signal();
-
-  static Widget _noteItem(BuildContext context, Value<To?> selected, ToNote node, RouteContext route, bool includeDraft) {
-    final colors = Theme.of(context).colorScheme;
-    final rootNoteLevel = routes.routes_notes.level + 1;
-
-    click() {
-      if (node.isLeafPage) {
-        route.to(node.toUri());
-        selected.value = node;
-      } else {
-        node.expand = !node.expand;
-      }
-    }
-
-    // 🔹◽️●○◦■□❏✎
-    String iconExtend = node.isLeafPage
-        ? "❏"
-        : node.expand
-            ? "▼"
-            : "︎︎︎▶";
-
-    String title = "$iconExtend ${node.label}";
-    title = title.padLeft(((node.level - rootNoteLevel) * 2) + title.length);
-    String publishLabel = "";
-    if (node.isLeafPage) {
-      publishLabel = node.isPublish ? "(已发布)" : "(草稿)";
-    }
-    return ListTile(
-      title: Text('$title$publishLabel'),
-      minTileHeight: 36,
-      minVerticalPadding: 6,
-      minLeadingWidth: 0,
-      selected: selected.value == node,
-      selectedTileColor: colors.surfaceContainerHighest,
-      onTap: click,
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    routes.routes_root.expandTree(true);
-    return Watch((context) {
-      final route = context.route$;
-      final colors = Theme.of(context).colorScheme;
-
-      final notes = routes.routes_notes.toList(includeThis: false).cast<ToNote>().where((e) {
-        return e.containsPage() && e.parent.expand && (includeDraft.value || e.containsPublishNode(includeThis: true));
-      });
-      return Column(
-        children: [
-          Container(
-            color: colors.surfaceContainerHighest,
-            child: OverflowBar(alignment: MainAxisAlignment.end, children: [
-              IconButton(tooltip: "Expand all", icon: const Icon(Icons.expand, size: 24), iconSize: 24, onPressed: () => notes.forEach((i) => i.expandTree(true))),
-              IconButton(tooltip: "Collapse all", icon: const Icon(Icons.compress), iconSize: 24, onPressed: () => notes.forEach((i) => i.expandTree(false))),
-              IconButton(tooltip: "Include draft", icon: const Icon(Icons.drafts_outlined), iconSize: 24, selectedIcon: const Icon(Icons.drafts), isSelected: includeDraft.value, onPressed: () => includeDraft.value = !includeDraft.value),
-              IconButton(tooltip: "Hidden", icon: const Icon(Icons.horizontal_rule), iconSize: 24, onPressed: () => widget.view.value = ""),
-            ]),
-          ),
-          const Divider(height: 1),
-          Material(
-            child: ListView(
-              padding: EdgeInsets.zero,
-              children: [...notes.map((e) => _noteItem(context, selected, e, route, includeDraft.value)).toList()],
-            ),
-          ).expanded$(),
-        ],
-      ).constrainedBox$(
-        constraints: const BoxConstraints.tightFor(width: 280),
-      );
-    });
-  }
-}
-
-extension _NoteTreeNode on To {
-  static Map<To, bool> expands = <To, bool>{}.signal();
-
-  bool get expand {
-    var result = expands[this];
-    return result ?? false;
-  }
-
-  set expand(bool value) {
-    expands[this] = value;
-  }
-
-  /// 展开层级数
-  /// level < 0 ,expand all levels
-  void expandTree(bool value, {int level = -1}) {
-    if (level == 0) return;
-
-    var nextLevel = level - 1;
-    expand = value;
-
-    if (nextLevel == 0) {
-      return;
-    }
-
-    for (var e in children) {
-      e.expandTree(value, level: nextLevel);
-    }
-  }
-}
-
 Future<void> _launchUrl(Uri url) async {
   if (!await url_launcher.launchUrl(url)) {
     throw Exception('Could not launch $url');
@@ -246,7 +128,10 @@ class ViewBarState extends State<ViewBar> {
               ),
               const VerticalDivider(width: 1),
               Watch((context) {
-                return _NoteTreeView(view: view).offstage$(offstage: view.value != "note_tree_view");
+                return NoteTreeView(
+                  view: view,
+                  noteRoot: routes.routes_notes,
+                ).offstage$(offstage: view.value != "note_tree_view");
               }),
               Watch((context) {
                 return _ThemeView(view: view).offstage$(offstage: view.value != "theme_view");
@@ -338,7 +223,7 @@ class _ThemeViewState extends State<_ThemeView> {
         ).paddingAll$(10),
         FilledButton(
             onPressed: () {
-              route.to(routes.routes_notes_style_theming_colors.toUri());
+              route.to(routes.routes_notes_material3_color_roles.toUri());
             },
             child: const Text("open Material 3 color roles page")),
       ],
