+++ packages/you_flutter/lib/note.dart
@@ -10,3 +10,5 @@ export 'src/note/contents/mockup.dart' show MockupWindow;
 export 'src/note/contents/markdown_content.dart' show MD;
 
 export 'src/note/visualization/material_colors.dart' show MaterialColorRoles;
+
+export 'src/note/note_tree_view.dart' show NoteTreeView;
